requirements.txt
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
dlib==20.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fonttools==4.61.1
joblib==1.5.3
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.1.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
scikit-learn==1.8.0
scipy==1.17.0
setuptools==80.9.0
six==1.17.0
threadpoolctl==3.6.0
